fix!: non-helpful max inlines error messages
Refactor derivation of Arbitrary-instances so that it is (much) more
likely that in case of "Maximal number of successive inlines exceeded",
the compiler error message will be the expected one.

This change is somewhat breaking as it increases the inlining steps for
products by one, so it slightly reduces the threshold for hitting that
limit. (This could have been avoided at the price of writing redundant
code, but that didn't seem worth it.)