fix!: avoid stack overflow (#106)
Use Gen.sized / Gen.resize to abort generation when recursion depth reaches the configured size.

BREAKING CHANGE: this is a runtime-breaking change as it changes the way how instances for sum types (sealed traits & enums) are derived. Hence, reproducing failed tests that ran with previous versions of this library using this new version of this library may not be possible.